Add Ti.Filesystem.File #createdAt() and #modifiedAt() methods that return Dates. Deprecate #modificationTimestamp() and #createTimestamp() due to inconsistent return types across platforms. Update unit tests to check new createdAt/modifiedAt methods. Add docs for new methods/deprecations. Use longs in place of doubles in Android file APIs whne it's more appropriate. Support getting creation timestamp for files on Android when API 26+. Fix spaceAvailable to use newer more accurate APIs on API level 18+.

+++ android/titanium/src/java/org/appcelerator/titanium/io/TiResourceFile.java
@@ -183,9 +183,9 @@ public class TiResourceFile extends TiBaseFile
 	}
 
 	@Override
-	public double spaceAvailable()
+	public long spaceAvailable()
 	{
-		return 0;
+		return 0L;
 	}
 
 	public String toURL()
@@ -196,7 +196,7 @@ public class TiResourceFile extends TiBaseFile
 	public long size()
 	{
 		if (!isFile()) {
-			return 0;
+			return 0L;
 		}
 
 		InputStream is = null;
@@ -214,7 +214,7 @@ public class TiResourceFile extends TiBaseFile
 				}
 			}
 		}
-		return 0;
+		return 0L;
 	}
 
 	@Override
